+++ server/jetstream_cluster_2_test.go
@@ -7260,85 +7260,7 @@ func TestJetStreamClusterCompressedStreamMessages(t *testing.T) {
 	}
 }
 
-func TestJetStreamClusterRemovePeerByID(t *testing.T) {
-	c := createJetStreamClusterExplicit(t, "R3S", 3)
-	defer c.shutdown()
-
-	s := c.randomNonLeader()
-	nc, js := jsClientConnect(t, s)
-	defer nc.Close()
-
-	_, err := js.AddStream(&nats.StreamConfig{
-		Name:     "TEST",
-		Subjects: []string{"foo", "bar"},
-		Replicas: 3,
-	})
-	require_NoError(t, err)
-
-	// Wait for a leader
-	c.waitOnStreamLeader(globalAccountName, "TEST")
-
-	// Get the name of the one that is not restarted
-	srvName := c.opts[2].ServerName
-	// And its node ID
-	peerID := c.servers[2].Node()
-
-	nc.Close()
-	// Now stop the whole cluster
-	c.stopAll()
-	// Restart all but one
-	for i := 0; i < 2; i++ {
-		opts := c.opts[i]
-		s, o := RunServerWithConfig(opts.ConfigFile)
-		c.servers[i] = s
-		c.opts[i] = o
-	}
-
-	c.waitOnClusterReadyWithNumPeers(2)
-	c.waitOnStreamLeader(globalAccountName, "TEST")
-
-	// Now attempt to remove by name, this should fail because the cluster
-	// was restarted and names are not persisted.
-	ml := c.leader()
-	nc, err = nats.Connect(ml.ClientURL(), nats.UserInfo("admin", "s3cr3t!"))
-	require_NoError(t, err)
-	defer nc.Close()
-
-	req := &JSApiMetaServerRemoveRequest{Server: srvName}
-	jsreq, err := json.Marshal(req)
-	require_NoError(t, err)
-	rmsg, err := nc.Request(JSApiRemoveServer, jsreq, 2*time.Second)
-	require_NoError(t, err)
-
-	var resp JSApiMetaServerRemoveResponse
-	err = json.Unmarshal(rmsg.Data, &resp)
-	require_NoError(t, err)
-	require_True(t, resp.Error != nil)
-	require_True(t, IsNatsErr(resp.Error, JSClusterServerNotMemberErr))
-
-	// Now try by ID, but first with an ID that does not match any peerID
-	req.Peer = "some_bad_id"
-	jsreq, err = json.Marshal(req)
-	require_NoError(t, err)
-	rmsg, err = nc.Request(JSApiRemoveServer, jsreq, 2*time.Second)
-	require_NoError(t, err)
-
-	resp = JSApiMetaServerRemoveResponse{}
-	err = json.Unmarshal(rmsg.Data, &resp)
-	require_NoError(t, err)
-	require_True(t, resp.Error != nil)
-	require_True(t, IsNatsErr(resp.Error, JSClusterServerNotMemberErr))
-
-	// Now with the proper peer ID
-	req.Peer = peerID
-	jsreq, err = json.Marshal(req)
-	require_NoError(t, err)
-	rmsg, err = nc.Request(JSApiRemoveServer, jsreq, 2*time.Second)
-	require_NoError(t, err)
-
-	resp = JSApiMetaServerRemoveResponse{}
-	err = json.Unmarshal(rmsg.Data, &resp)
-	require_NoError(t, err)
-	require_True(t, resp.Error == nil)
-	require_True(t, resp.Success)
-}
+//
+// DO NOT ADD NEW TESTS IN THIS FILE
+// Add at the end of jetstream_cluster_<n>_test.go, with <n> being the highest value.
+//
